backend/requirements.txt
APScheduler==3.8.1
asgiref==3.3.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.9
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-model-utils==4.2.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
drf-extra-fields==3.2.1
gunicorn==20.0.4
idna==3.3
importlib-resources==5.4.0
langid==1.1.6
mysqlclient==2.1.0
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
psycopg2==2.7.7
psycopg2-binary==2.9.3
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
